refactor: extract readLoggedInCustomer into shared utils/auth.js
Agent-Logs-Url: https://github.com/BSPQ25-26/BSPQ26-E5/sessions/a2d9d118-2e1b-445c-9818-7c6705ee358f
